# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Define Pilates API",
        "description": "API documentation for Define Pilates management system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://sibilation-edythe-unprivately.ngrok-free.dev",
            "description": "Ngrok Server (HTTPS) - Use when accessing via Ngrok"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Server (HTTP) - Use when accessing locally"
        }
    ],
    "paths": {
...
        "/api/admin_receptionist/save": {
            "post": {
                "tags": [
                    "AdminReceptionist"
                ],
                "summary": "Cadastro completo de novo administrador ou recepcionista",
                "operationId": "registerAdminReceptionist",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminReceptionistRegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminReceptionistSuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro interno"
                    }
                }
            }
        },
        "/api/admin_receptionist/login": {
            "post": {
                "tags": [
                    "AdminReceptionist"
                ],
                "summary": "Login de administrador ou recepcionista",
                "operationId": "loginAdminReceptionist",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/students/save": {
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Cadastro completo de novo estudante",
                "operationId": "registerStudent",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StudentRegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StudentSuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro interno"
                    }
                }
            }
        },
        "/api/students/list": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Listar todos os estudantes com informações básicas",
                "operationId": "listStudents",
                "responses": {
                    "200": {
                        "description": "Lista de estudantes retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "fullname": {
                                                        "type": "string",
                                                        "example": "João Silva"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "11999999999"
                                                    },
                                                    "typeuser": {
                                                        "type": "string",
                                                        "example": "student"
                                                    },
                                                    "birth_date": {
                                                        "type": "string",
                                                        "example": "1990-01-15",
                                                        "nullable": true
                                                    },
                                                    "cpf": {
                                                        "type": "string",
                                                        "example": "12345678901",
                                                        "nullable": true
                                                    },
                                                    "profession": {
                                                        "type": "string",
                                                        "example": "Engenheiro",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao buscar estudantes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to list students"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/students/login": {
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Login de estudante",
                "operationId": "loginStudent",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AdminReceptionistRegisterRequest": {
                "title": "Admin/Receptionist Register Request",
                "description": "Request body for admin or receptionist registration",
                "required": [
                    "name",
                    "phone",
                    "password",
                    "birth_date",
                    "hiring",
                    "classification",
                    "typecollaborator",
                    "fulladdress"
                ],
                "properties": {
                    "name": {
                        "description": "Nome completo do colaborador",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Maria Santos"
                    },
                    "phone": {
                        "description": "Telefone principal (formato: 11999999999 ou (11)99999-9999)",
                        "type": "string",
                        "example": "(11)98765-4321"
                    },
                    "password": {
                        "description": "Senha de acesso (mínimo 8 caracteres, deve conter maiúscula, minúscula, número e caractere especial)",
                        "type": "string",
                        "maxLength": 16,
                        "minLength": 8,
                        "example": "Senha@123"
                    },
                    "birth_date": {
                        "description": "Data de nascimento (formato DD-MM-YYYY)",
                        "type": "string",
                        "format": "date",
                        "example": "15-03-1985"
                    },
                    "hiring": {
                        "description": "Data de contratação (formato DD-MM-YYYY)",
                        "type": "string",
                        "format": "date",
                        "example": "01-02-2024"
                    },
                    "classification": {
                        "description": "Classificação profissional (1 letra)",
                        "type": "string",
                        "maxLength": 1,
                        "example": "A"
                    },
                    "typecollaborator": {
                        "description": "Tipo de colaborador",
                        "type": "string",
                        "enum": [
                            "Administrator",
                            "Receptionist"
                        ],
                        "example": "Administrator"
                    },
                    "fulladdress": {
                        "description": "Endereço completo (mínimo 15 caracteres, máximo 255)",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 15,
                        "example": "Rua XV de Novembro, 745, Centro, Curitiba - PR"
                    }
                },
                "type": "object"
            },
            "AdminReceptionistSuccessResponse": {
                "title": "Admin/Receptionist Success Response",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Admin/Receptionist registered successfully"
                    },
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Maria Santos"
                            },
                            "phone": {
                                "type": "string",
                                "example": "(11)98765-4321"
                            },
                            "typecollaborator": {
                                "type": "string",
                                "example": "Administrator"
                            },
                            "classification": {
                                "type": "string",
                                "example": "A"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Contato": {
                "title": "Contact Information",
                "required": [
                    "tipo",
                    "valor"
                ],
                "properties": {
                    "tipo": {
                        "type": "string",
                        "enum": [
                            "telefone",
                            "email",
                            "whatsapp",
                            "emergencia"
                        ],
                        "example": "email"
                    },
                    "valor": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "joao@email.com"
                    },
                    "observacao": {
                        "type": "string",
                        "maxLength": 500,
                        "example": "Email pessoal",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Endereco": {
                "title": "Address Information",
                "required": [
                    "tipo",
                    "rua",
                    "numero",
                    "bairro",
                    "cidade",
                    "estado",
                    "cep"
                ],
                "properties": {
                    "tipo": {
                        "type": "string",
                        "enum": [
                            "residencial",
                            "comercial",
                            "cobranca"
                        ]
                    },
                    "rua": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Rua das Flores"
                    },
                    "numero": {
                        "type": "string",
                        "maxLength": 10,
                        "example": "123"
                    },
                    "complemento": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Apto 45",
                        "nullable": true
                    },
                    "bairro": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "Centro"
                    },
                    "cidade": {
                        "type": "string",
                        "maxLength": 100,
                        "example": "São Paulo"
                    },
                    "estado": {
                        "type": "string",
                        "maxLength": 2,
                        "example": "SP"
                    },
                    "cep": {
                        "type": "string",
                        "example": "01234-567"
                    },
                    "principal": {
                        "type": "boolean",
                        "example": true,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "LoginErrorResponse": {
                "title": "Login Error Response",
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Credenciais inválidas"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "title": "Login Request",
                "required": [
                    "nameuser",
                    "password"
                ],
                "properties": {
                    "login": {
                        "description": "Nome de usuário (Telefone)",
                        "type": "string",
                        "example": "11951999999"
                    },
                    "password": {
                        "description": "Senha do usuário",
                        "type": "string",
                        "format": "password",
                        "example": "abc123A"
                    }
                },
                "type": "object"
            },
            "LoginSuccessResponse": {
                "title": "Login Success Response",
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Login realizado com sucesso"
                    },
                    "data": {
                        "properties": {
                            "token": {
                                "description": "Token de autenticação JWT",
                                "type": "string",
                                "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                            },
                            "student": {
                                "properties": {
                                    "Id_students": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "11951999999"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "profession": {
                                        "type": "string",
                                        "example": "Engenheiro"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "15-01-1990"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "StudentRegisterRequest": {
                "title": "Student Register Request",
                "description": "Request body for student registration",
                "required": [
                    "name",
                    "phone",
                    "password",
                    "cpf",
                    "profession",
                    "birth_date"
                ],
                "properties": {
                    "name": {
                        "description": "Nome completo do estudante",
                        "type": "string",
                        "maxLength": 255,
                        "example": "João Silva"
                    },
                    "phone": {
                        "description": "Telefone principal (usado como login)",
                        "type": "string",
                        "example": "(11)99999-9999"
                    },
                    "password": {
                        "description": "Senha de acesso (mínimo 6 caracteres)",
                        "type": "string",
                        "minLength": 6,
                        "example": "abc123A"
                    },
                    "cpf": {
                        "description": "CPF do estudante (apenas números)",
                        "type": "string",
                        "maxLength": 11,
                        "minLength": 11,
                        "example": "12345678901"
                    },
                    "profession": {
                        "description": "Profissão do estudante",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Engenheiro"
                    },
                    "birth_date": {
                        "description": "Data de nascimento (formato DD-MM-YYYY)",
                        "type": "string",
                        "format": "date",
                        "example": "15-01-1990"
                    },
                    "fotos": {
                        "description": "Fotos do estudante em Base64 (máximo 5 fotos, 3MB cada)",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "base64",
                            "example": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD..."
                        },
                        "maxItems": 5,
                        "nullable": true
                    },
                    "contatos": {
                        "description": "Contatos adicionais do estudante (máximo 3)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Contato"
                        },
                        "maxItems": 3,
                        "nullable": true
                    },
                    "enderecos": {
                        "description": "Endereços do estudante (máximo 2)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Endereco"
                        },
                        "maxItems": 2,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "StudentSuccessResponse": {
                "title": "Success Response",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Student registered successfully"
                    },
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "João Silva"
                            },
                            "phone": {
                                "type": "string",
                                "example": "(11)99999-9999"
                            },
                            "has_photos": {
                                "type": "boolean",
                                "example": true
                            },
                            "has_additional_data": {
                                "type": "boolean",
                                "example": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}